--- services/newsfeed/shared/drupal.js
@@ -1,37 +1,0 @@
-/*
- * LiskHQ/lisk-service
- * Copyright © 2021 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-const { HTTP } = require('lisk-service-framework');
-
-const requestLib = HTTP.request;
-const { normalizeData } = require('./normalizers');
-
-const mysqlIndex = require('./indexdb/mysql');
-const newsfeedIndexSchema = require('./schema/newsfeed');
-
-const getnewsfeedIndex = () => mysqlIndex('newsfeed', newsfeedIndexSchema);
-
-const reloadNewsFromDrupal = async drupalSources => {
-	const newsfeedDB = await getnewsfeedIndex();
-	drupalSources.forEach(async source => {
-		const response = await requestLib(source.url);
-		const normalizedData = normalizeData(source, response.data);
-		await newsfeedDB.upsert(normalizedData);
-	});
-};
-
-module.exports = {
-	reloadNewsFromDrupal,
-};
